hardware/riscv: Fix combinatorial loop and FPU latency in ID stage
The loop was caused by the `stall_full` signal in the `apu_dispatcher`.
It has been removed since with the new changes to the module it was
never SET.

Also modified `csr_apu_stall` signal in the `riscv_id_stage` to
be consistent with the new latency encoding for FPU operations.