support: Update IRQ lanes on ZCU102
On the ZCU102, we no longer have an external interrupt controller to
aggregate IRQs from PULP, hence the `INTR_REG_BASE_ADDR` address is set
to `0`.  This means multiple IRQ lanes from PULP are now directly
connected to the processor.  The reason for this is that the driver of
the interrupt controller seems to interfere with our own IRQ handling.

Additionally, most of the IRQs from RAB are split into Host and PULP.
This has been the case for a longer time, but it had not been reflected
in the driver.