board/xilzcu102: Increase size of boot partition to 256 MiB
This is required because the slightly larger root FS no longer fits into
the previous limit of 128 MiB.  256 MiB should provide ample space and
also allows to add some extra configurations in the future or to
experiment in existing deployments.  Larger sizes currently fail with an
obscure error ("Total number of sectors (..) not a multiple of sectors
per track (63)!").

I think not all of the files that we currently add to the boot partition
are required, though.  The minimum should be the `image.ub`, containing
the Linux kernel image, the device tree blob, and the root FS image, and
the boot.bin, which should contain the FSBL, the U-boot boot loader, the
FPGA bitstream, the PMU FW, and the ATF FW.