support/libpulp: Protect L3 against buffer overflows
This adds a known value ("canary") at the start of the `Fragment`
`struct` in o1heap.  The `Fragment` stores memory allocation meta data
and it is located in memory after user-allocated data.  Thus, when a
buffer in the user-allocated memory overflows and corrupts the
`Fragment`, it corrupts the canary value first.  This will trigger an
assertion and abort execution.