apps/physmem: Make `write` functions non-`const`
Rationale: While these member functions do not modify the actual object
(in a strict C++ sense), they write to the underlying physical memory
region, which is represented by object.  In this sense, the `write`
functions should not be `const`, so one can be sure that a `const
PhysMem` object is never written to.