pulp-driver (pulp_mmap): Replace `remap_pfn_range` with `io_remap_pfn_range`
According to the latest Linux Device Drivers Development book [1], "the
`remap_pfn_range()` function [..] does not apply anymore when it comes
to mapping I/O memory to user space.  In this case, the appropriate
function is `io_remap_pfn_range()`, whose parameters are the same."

If this is true, our use of `remap_pfn_range()` instead of
`io_remap_pfn_range()` could have led the kernel to optimize or
transform accesses to the mapped I/O memory in a way that does not
correspond to exact accesses to device memory.  Thereby, this could be a
cause for #87.

I briefly tried to reinforce this suspicion with the kernel code
implementing `io_remap_pfn_range()` [2].  However, the search [2] does
not show a specialized implementation for ARM, and the default
implementation [3] simply calls `remap_pfn_range()`.  The only
difference is that the `prot` argument is wrapped with
`pgprot_decrypted()`, which does not have a specialized implementation
for ARM [4] and in its default implementation is a no-op [5].

To conclude, I did not find code evidence that would back that using
`io_remap_pfn_range()` instead of `remap_pfn_range()` actually causes
issue #87, but since we are dealing with I/O memory, it anyway seems to
be the correct function to use here.

[1]: https://www.oreilly.com/library/view/linux-device-drivers/9781785280009/0cf09217-f38e-4c5e-9123-a24d1cbf01d2.xhtml
[2]: https://elixir.bootlin.com/linux/v4.19.165/A/ident/io_remap_pfn_range
[3]: https://elixir.bootlin.com/linux/v4.19.165/source/include/linux/mm.h#L2565
[4]: https://elixir.bootlin.com/linux/v4.19.165/C/ident/pgprot_decrypted
[5]: https://elixir.bootlin.com/linux/v4.19.165/source/include/asm-generic/pgtable.h#L638