setup.sh: Fix invocation of `make`
Invoking `make` with `-C` causes it to enable printing of directories.
This can be useful for debugging recursive invocations of `make`, but
for some reasons the "Entering directory [...]" and "Leaving directory
[...]" messages clutter the standard output of commands invoked within
targets.  In our case, this causes `ct-ng -s show-tuple` to return the
tuple with an "Entering directory" message before and a "Leaving
directory" message after the tuple.  In that way, the tuple cannot be
used for further processing.