support/libhero-target/pulp: Fix pointer width mismatch in `hero_memcpy_*_async`
These functions differ only in the address spaces (ASes) of their
arguments.  Each AS has a specific address width. Thus, pointers must
not be casted to `void*` because pointers from ASes wider than the
default AS will then lose address bits.  Instead, each pointer must be
cast to an integer of corresponding width.  This commit fixes that.

Currently, both device and host addresses are casted to 32-bit
integers, but this will be fixed separately soon.