openmp-examples/tests-pulp/perf: Fix comparison for retired instructions
The prior number of instructions in the equality comparison was just a
first guess to get an actual result by measurement.  Now that the
measurement was verified, the number has been updated.

Moreover, the number of instructions might vary in the future, and we do
not want this test to fail if it slightly decreases or increases.  We
thus now compare with a margin of 10% to an expected value.  Thus, if
there is a significant increase in the number of instructions measured
in this test, the test will still fail (which may indicate a performance
regression).  Additionally, the test will fail on a significant
*decrease* in the number of instructions: in that case, we need to
update the expected value as a regression test for the improved
performance.