openmp-examples/darknet-layer: Compile `gemm_layers.c` independently
Prior to this commit, `gemm_layers.c` was `#include`d from `darknet` by
`gemm.c`, which is not the usual way of organizing C compilation units.
This commit symlinks `darknet/gemm_layers.c` (and the `gemm.h`, which is
included in `main.c`) and adds `gemm_layers.c` as separate compilation
unit.