openmp-examples/darknet: Compile `gemm_layers.c` independently
Prior to this commit, `gemm_layers.c` was `#include`d by `gemm.c`,
which is not the usual way of organizing C compilation units.
The reason might be that `gemm_layers.c` missed two symbols:
`LAYER_COUNTER`, declared within `gemm.c` and `BIGPULP_MEMCPY` from
`libhero-target`.  With the former declared and an include for the
header of the latter added, `gemm_layers.c` now compiles to its own
object file.