hardware/riscv: In simulation, count all HPM events and print them at the end
This feature is disabled by default.  To activate it, change
`SIM_COUNT_HPMEVENTS` to `1'b1`.

QuestaSim must run in console mode to print the messages of the `final`
statement at the end of simulation.  In GUI mode, the messages are not
shown (at least not by default, maybe there is an option to activate
them).