openmp-examples: Re-add `pulp.dis` rule
This rule extracts the device offloading image from the heterogeneous
(fat) binary and disassembles it with the device toolchain.  This is
required for introspection and debugging of the code running on the
device (PULP in this case).

This commit also reverts the CI changes made in the commit with message

    "openmp-examples: Removed pulp.dis rule, as LLVM12 has new FAT binary format"

when the `pulp.dis` rule was removed.